move Node to it's own files
